Fix upload success modal and file count display
- Fix system_events table ID generation in DuckDB (manual ID generation)
- Remove delays from upload success modal to show immediately
- Fix file count in success modal to correctly show multiple files uploaded
- Add date picker placeholder text and future date restriction
- Improve batch upload state tracking for accurate file counts
- Update UploadSuccessModal to trigger animation immediately

+++ backend/core/database.py
@@ -264,9 +264,10 @@ def init_database():
             pass
         
         # Create system_events table for monitoring/observability
+        # DuckDB INTEGER PRIMARY KEY doesn't auto-increment, so we'll use manual ID generation
         conn.execute("""
             CREATE TABLE IF NOT EXISTS system_events (
-                id INTEGER PRIMARY KEY,
+                id BIGINT PRIMARY KEY,
                 created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
                 level VARCHAR NOT NULL,
                 category VARCHAR,
